requirements.txt
annotated-types==0.7.0
anyio==4.11.0
apimatic-core==0.2.23
apimatic-core-interfaces==0.1.8
apimatic-requests-client-adapter==0.1.9
asgiref==3.11.0
beautifulsoup4==4.14.2
CacheControl==0.12.14
certifi==2025.11.12
charset-normalizer==3.4.4
deprecation==2.1.0
Django==4.0.10
django-anymail==11.1
django-bootstrap5
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpickle==3.3.0
jsonpointer==2.4
msgpack==1.1.2
mysqlclient==2.2.7
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
squareup==43.2.0.20251016
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
